navigation-api/navigate-event/signal-abort-window-stop.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../navigation-methods/return-value/resources/helpers.js"></script>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));
  let abort_signal;
  let onabort_called = false;
  navigation.onnavigatesuccess = t.unreached_func("onnavigatesuccess");
  navigation.onnavigate = t.step_func(e => {
    abort_signal = e.signal;
    abort_signal.onabort = () => onabort_called = true;
  });
  const navigationToStop = navigation.navigate("?1");
  window.stop();
  await assertBothRejectDOM(t, navigationToStop, "AbortError");
  assert_true(abort_signal.aborted);
  assert_true(onabort_called);
  // Complete one final task to ensure that onnavigatesuccess didn't fire on a
  // microtask.
  await new Promise(resolve => t.step_timeout(resolve, 0));
}, "window.stop() signals event.signal");
</script>
